moved tela.py to roll.py and fixed it
